+++ src/main/java/de/stklcode/jvault/connector/model/response/embedded/SecretMetadata.java
@@ -58,19 +58,6 @@ public final class SecretMetadata implements Serializable {
     private Map<Integer, VersionMetadata> versions;
 
     /**
-     * @return Time of secret creation as raw string representation.
-     * @deprecated Method left for backwards compatibility only. Use {@link #getCreatedTime()} instead.
-     */
-    @Deprecated(since = "1.2", forRemoval = true)
-    public String getCreatedTimeString() {
-        if (createdTime != null) {
-            return TIME_FORMAT.format(createdTime);
-        }
-
-        return null;
-    }
-
-    /**
      * @return Time of secret creation.
      */
     public ZonedDateTime getCreatedTime() {
@@ -96,19 +83,6 @@ public final class SecretMetadata implements Serializable {
      */
     public Integer getOldestVersion() {
         return oldestVersion;
-    }
-
-    /**
-     * @return Time of secret update as raw string representation.
-     * @deprecated Method left for backwards compatibility only. Use {@link #getUpdatedTime()} instead.
-     */
-    @Deprecated(since = "1.2", forRemoval = true)
-    public String getUpdatedTimeString() {
-        if (updatedTime != null) {
-            return TIME_FORMAT.format(updatedTime);
-        }
-
-        return null;
     }
 
     /**